Use same types as in inierr from where the line and pos is copied.

+++ src/cxx/inifile++.hpp
@@ -88,8 +88,8 @@ namespace inifilepp {
 		//
 		struct exception
 		{
-			unsigned int line;
-			unsigned int cpos;
+			size_t line;
+			size_t cpos;
 			std::string msg;
 			exception(inifile *, const inierr *);
 		};
